+++ test/dog/http_dog_test.dart
@@ -2,26 +2,28 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:vexana/vexana.dart';
 
 // ignore: always_declare_return_types
-main() {
+void main() {
   late INetworkManager networkManager;
   setUp(() {
     networkManager = NetworkManager<EmptyModel>(
-        isEnableLogger: true,
-        interceptor: InterceptorsWrapper(
-          onRequest: (options, handler) {
-            print(options.data);
-            handler.next(options);
-          },
-        ),
-        options: BaseOptions(
-          baseUrl: 'https://hwasampleapi.firebaseio.com',
-        ));
+      isEnableLogger: true,
+      interceptor: InterceptorsWrapper(
+        onRequest: (options, handler) {
+          print(options.data);
+          handler.next(options);
+        },
+      ),
+      options: BaseOptions(
+        baseUrl: 'https://hwasampleapi.firebaseio.com',
+      ),
+    );
   });
   test('Primitve Type', () async {
     final response = await networkManager.send<EmptyModel, EmptyModel>(
-        '/dogs/0/code.json',
-        parseModel: EmptyModel(),
-        method: RequestType.GET);
+      '/dogs/0/code.json',
+      parseModel: EmptyModel(),
+      method: RequestType.GET,
+    );
 
     expect(response.data, isNotNull);
   });