+++ test/file-download/file_download_test.dart
@@ -6,15 +6,17 @@ import 'package:vexana/vexana.dart';
 import 'file_download_model.dart';
 
 // ignore: always_declare_return_types
-main() {
+void main() {
   late INetworkManager networkManager;
   setUp(() {
-    networkManager = NetworkManager<Null>(
-        isEnableLogger: true, options: BaseOptions(baseUrl: 'http://www.africau.edu/images/default/'));
+    networkManager = NetworkManager<EmptyModel>(
+      isEnableLogger: true,
+      options: BaseOptions(baseUrl: 'http://www.africau.edu/images/default/'),
+    );
   });
   test('Primitive Type', () async {
-    final response =
-        await networkManager.downloadFileSimple('http://www.africau.edu/images/default/sample.pdf', (count, total) {
+    final response = await networkManager.downloadFileSimple(
+        'http://www.africau.edu/images/default/sample.pdf', (count, total) {
       print('${count}');
     });
     expect(response.data, isList);