+++ test/refresh-token/refresh_user_token_test.dart
@@ -4,24 +4,28 @@ import 'package:vexana/vexana.dart';
 import 'user.dart';
 
 // ignore: always_declare_return_types
-main() {
-  late INetworkManager networkManager;
+void main() {
+  late INetworkManager<EmptyModel> networkManager;
   setUp(() {
-    networkManager = NetworkManager<Null>(
-        isEnableLogger: true,
-        // onRefreshFail: () {  Navigate to login },
-        onRefreshToken: (error, newService) async {
-          // final refreshToken = await newService.send<Credantial, Credantial>("/token",
-          //     parseModel: Credantial(), method: RequestType.GET);
-          // error.request.headers = {HttpHeaders.authorizationHeader: "Bearer ${refreshToken.data.token}"};
-          return error;
-        },
-        options: BaseOptions(baseUrl: 'http://localhost:3000'));
+    networkManager = NetworkManager<EmptyModel>(
+      isEnableLogger: true,
+      // onRefreshFail: () {  Navigate to login },
+      onRefreshToken: (error, newService) async {
+        // final refreshToken = await newService.send<Credantial, Credantial>("/token",
+        //     parseModel: Credantial(), method: RequestType.GET);
+        // error.request.headers = {HttpHeaders.authorizationHeader: "Bearer ${refreshToken.data.token}"};
+        return error;
+      },
+      options: BaseOptions(baseUrl: 'http://localhost:3000'),
+    );
   });
 
   test('Json Place Holder Todos', () async {
-    final response = await networkManager.send<User, User>('/user',
-        parseModel: User(), method: RequestType.GET);
+    final response = await networkManager.send<User, User>(
+      '/user',
+      parseModel: User(),
+      method: RequestType.GET,
+    );
 
     expect(response.data, isNotNull);
   });